+++ src/lib.rs
@@ -86,6 +86,7 @@ pub mod decoder;
 pub mod encode;
 pub mod encoder;
 pub mod header;
+pub mod model2;
 pub mod psy;
 pub mod synthesis;
 pub mod tables;